tests: remove unsupported use case planned on the test
This test was verifying the correctness of when a metric tool is called without files being passed explicitly which shouldmake the tool verify the whole src directory. But on the code flow this case never happens and we are always explicit in the files we want analyzed when sending them to the tools.