Remove test-docker-socket
A customer is trying to understand why we need to test this. He says:
"why do we need to test the socket file? Given the evolving nature of docker and K8s, the checking of socket file doesn’t seem applicable anymore."

More details in the ticket https://codacy.atlassian.net/browse/CY-6629
@machadoit @pedrocodacy @lolgab